+++ src/main/java/org/mmarini/wheelly/objectives/NoMove.java
@@ -27,8 +27,10 @@ package org.mmarini.wheelly.objectives;
 
 
 import com.fasterxml.jackson.databind.JsonNode;
-import org.eclipse.collections.api.block.function.primitive.FloatFunction;
+import org.eclipse.collections.api.block.function.primitive.DoubleFunction;
+import org.mmarini.rl.envs.Environment;
 import org.mmarini.wheelly.apis.RobotStatus;
+import org.mmarini.wheelly.envs.WithRobotStatus;
 import org.mmarini.yaml.schema.Locator;
 import org.mmarini.yaml.schema.Validator;
 
@@ -42,7 +44,7 @@ import static org.mmarini.yaml.schema.Validator.nonNegativeNumber;
  */
 public interface NoMove {
     float DEFAULT_VELOCITY_THRESHOLD = 0.01f;
-    FloatFunction<RobotStatus> NO_MOVE = noMove(DEFAULT_VELOCITY_THRESHOLD);
+    DoubleFunction<Environment> NO_MOVE = noMove(DEFAULT_VELOCITY_THRESHOLD);
     Validator VALIDATOR = Validator.objectProperties(Map.of(
             "velocityThreshold", nonNegativeNumber()
     ));
@@ -53,7 +55,7 @@ public interface NoMove {
      * @param root    the root json document
      * @param locator the locator
      */
-    static FloatFunction<RobotStatus> create(JsonNode root, Locator locator) {
+    static DoubleFunction<Environment> create(JsonNode root, Locator locator) {
         VALIDATOR.apply(locator).accept(root);
         float velocityThreshold = (float) locator.path("velocityThreshold").getNode(root).asDouble(DEFAULT_VELOCITY_THRESHOLD);
         return noMove(velocityThreshold);
@@ -62,7 +64,7 @@ public interface NoMove {
     /**
      * Returns the function that rewards the no move behavior
      */
-    static FloatFunction<RobotStatus> noMove() {
+    static DoubleFunction<Environment> noMove() {
         return NO_MOVE;
     }
 
@@ -71,13 +73,16 @@ public interface NoMove {
      *
      * @param velocityThreshold the velocity threshold
      */
-    static FloatFunction<RobotStatus> noMove(float velocityThreshold) {
-        return status -> !status.canMoveForward() || !status.canMoveBackward()
-                ? -1
-                : (abs(status.getLeftPps()) < velocityThreshold
-                && abs(status.getRightPps()) < velocityThreshold
-                && status.getSensorDirection() == 0) ? 1
-                : 0;
+    static DoubleFunction<Environment> noMove(float velocityThreshold) {
+        return environment -> {
+            RobotStatus status = ((WithRobotStatus) environment).getStatus();
+            return !status.canMoveForward() || !status.canMoveBackward()
+                    ? -1
+                    : (abs(status.getLeftPps()) < velocityThreshold
+                    && abs(status.getRightPps()) < velocityThreshold
+                    && status.getSensorDirection() == 0) ? 1
+                    : 0;
+        };
 
     }
 }